Limit input to one file
Previous behavior would allow multiple files as input and use only the
header from the first file to determine what columns to output, which
could lead to unexpected results if the header of the subsequent files
do not match. It doesn't make sense to have this added complexity
without any good use cases.